release 3.0.732

+++ lighthouse/include/tencentcloud/lighthouse/v20200324/model/PolicyDetail.h
@@ -50,13 +50,13 @@ namespace TencentCloud
                      * 获取用户折扣。
                      * @return UserDiscount 用户折扣。
                      */
-                    int64_t GetUserDiscount() const;
+                    double GetUserDiscount() const;
 
                     /**
                      * 设置用户折扣。
                      * @param UserDiscount 用户折扣。
                      */
-                    void SetUserDiscount(const int64_t& _userDiscount);
+                    void SetUserDiscount(const double& _userDiscount);
 
                     /**
                      * 判断参数 UserDiscount 是否已赋值
@@ -68,13 +68,13 @@ namespace TencentCloud
                      * 获取公共折扣。
                      * @return CommonDiscount 公共折扣。
                      */
-                    int64_t GetCommonDiscount() const;
+                    double GetCommonDiscount() const;
 
                     /**
                      * 设置公共折扣。
                      * @param CommonDiscount 公共折扣。
                      */
-                    void SetCommonDiscount(const int64_t& _commonDiscount);
+                    void SetCommonDiscount(const double& _commonDiscount);
 
                     /**
                      * 判断参数 CommonDiscount 是否已赋值
@@ -86,13 +86,13 @@ namespace TencentCloud
                      * 获取最终折扣。
                      * @return FinalDiscount 最终折扣。
                      */
-                    int64_t GetFinalDiscount() const;
+                    double GetFinalDiscount() const;
 
                     /**
                      * 设置最终折扣。
                      * @param FinalDiscount 最终折扣。
                      */
-                    void SetFinalDiscount(const int64_t& _finalDiscount);
+                    void SetFinalDiscount(const double& _finalDiscount);
 
                     /**
                      * 判断参数 FinalDiscount 是否已赋值
@@ -153,19 +153,19 @@ user：用户折扣; common：官网折扣; activity：活动折扣。 取值为
                     /**
                      * 用户折扣。
                      */
-                    int64_t m_userDiscount;
+                    double m_userDiscount;
                     bool m_userDiscountHasBeenSet;
 
                     /**
                      * 公共折扣。
                      */
-                    int64_t m_commonDiscount;
+                    double m_commonDiscount;
                     bool m_commonDiscountHasBeenSet;
 
                     /**
                      * 最终折扣。
                      */
-                    int64_t m_finalDiscount;
+                    double m_finalDiscount;
                     bool m_finalDiscountHasBeenSet;
 
                     /**
